Fix access levels in the inner interfaces and corresponding files. Move test console interface into presentation level.
